Issue #11163: Enforce file size on InputJavadocTypePublicOnly1.java
