+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadoctype/InputJavadocTypePublicOnly1.java
@@ -54,74 +54,4 @@ public class InputJavadocTypePublicOnly1
     int mLen;
     protected int mDeer;
     public int aFreddo;
-
-
-    private InputJavadocTypePublicOnly1(int aA)
-    {
-    }
-
-
-    InputJavadocTypePublicOnly1(String aA)
-    {
-    }
-
-
-    protected InputJavadocTypePublicOnly1(Object aA)
-    {
-    }
-
-
-    public InputJavadocTypePublicOnly1(Class<Object> aA)
-    {
-    }
-
-
-    private void method(int aA)
-    {
-    }
-
-
-    void method(Long aA)
-    {
-    }
-
-
-    protected void method(Class<Object> aA)
-    {
-    }
-
-
-    public void method(StringBuffer aA)
-    {
-    }
-
-
-    /**
-       A param tag should not be required here when relaxed about Javadoc.
-       Writing a little documentation should not be worse than not
-       writing any documentation at all.
-     */
-    private void method(String aA)
-    {
-    }
-
-    /**
-       This inner class has no author tag, which is OK.
-     */
-    public class InnerWithoutAuthor
-    {
-
-    }
-
-    /** {@inheritDoc} */
-    public String toString()
-    {
-        return super.toString();
-    }
-
-    @Deprecated @Override
-    public int hashCode()
-    {
-        return super.hashCode();
-    }
 }
